Issue 25: only bind timeago plugin once-per-element
This change will filter out any elements that have already been bound to
the timeago plugin so they are not included in multiple setInterval
closures.

NOTE: I also had to bump the version of jQuery used as 1.4.2 has a
bug where it wouldn't call the .each(callback). Perhaps this is due to
changes to the .filter method? Moving to 1.4.1 or 1.4.4
is also acceptable.