no more memory leaks caused by removing elements
Now whenever you call .timeago() it has a class added to it that allows for it to update every minute or half hour or not at all. This methodology removes memory leaks caused by elements being removed from the page (setIntervals calling functions for missing elements or events firing for missing elements). If you uncomment the lines $win.on('timeago') and the global timeout for timeago it will automatically hook any elements with the class .timeago every 250 ms. This new methodology removes having an event hook for each element, now there is only 2 (or 3 if you uncomment) timeouts and 2 (or 3 if you uncomment) events hooks, ever. It also updates the timeagos from min to hour or hour to never. if the preparedata functions (and istime) were all in a web worker this can probably be improved even more by leveraging multiple cores.