changed the words function to a "lookup table"
In JS there are a few ways to do lookup tables, this "table" decreases the average amount of functions performed for each kind of fuzzy time from 5.5 to 2.25. Meaning instead of the computer needing to process 11 statements to get to years, it will now only do 3. I also removed all the extraneous time calculations (it won't bother calculating days or hours unless it needs to display them). This should greatly increase the time it takes to build the fuzzy sentences in almost every possible case. For comparison http://jsperf.com/timeago-speed-test it seems to be about a 200% increase in speed.