Changed words var into a table lookup
Modified words var so it uses a table lookup to decrease cost to run timeago calculations. This should result in a notable speed increase. Instead of relying on a bunch of or statements where it would average 5.5 function calls per timeago, it now should now average 2.25. Also moved calculations of minute hour day and year so it only calculates them when needed.