Fix incorrect translation for "seconds"
according to the locale template, "seconds" means "less than a
minute". The term "1 dakikadan" is by itself meaningless. We
should say "1 dakikadan az bir süre" but that's IMHO a too long
phrase. My suggestion is to use "birkaç saniye", which means "a
few seconds". This phrase might also be techinally inconvenient
because of the fact that it's used for some duration less than
45 seconds, and saying "a few seconds" while 40 seconds have
passed might be... Weird. But not that bad.